pass network configuration in kernel command line
For a strange reasons, sometimes, anaconda/dracut fails back to dhcp.
My guess is that kickstart parsing might occure after dracut start the
network. Dracut is able to read the command line and recover network
parameters from there, so we configure all the network interfaces
directly in the sysconfig template.